airflow/requirements.txt
requests==2.32.3
scrapy==2.11.2
# ipython
python-dotenv==1.1.0
# psycopg2
# scrapy-deltafetch
google-cloud-storage==2.19.0
google-cloud-bigquery==3.31.0
apache-airflow-providers-google
dbt-bigquery==1.9.2
dbt-core==1.9.5